+++ lib/hive/Commands/BaseCommand.ts
@@ -1,8 +1,5 @@
-import { type } from 'os';
 import TCLIService from '../../../thrift/TCLIService';
-import HiveDriverError from '../../errors/HiveDriverError';
-import errorHandler from '../../utils/errorHandler';
-import { TApplicationException, TProtocolException } from '../..';
+import errorHandler from '../Utils/errorHandler';
 
 export default abstract class BaseCommand {
   protected client: TCLIService.Client;
@@ -12,6 +9,6 @@ export default abstract class BaseCommand {
   }
 
   executeCommand<Response>(request: object, command: Function | void): Promise<Response> {
-    return errorHandler<Response>(this.client, request, command, {numRetries: 0, startTime: Date.now()});
+    return errorHandler<Response>(this.client, request, command, { numRetries: 0, startTime: Date.now() });
   }
 }
