+++ tests/e2e/error_handling.test.js
